+++ malachite-nz/src/natural_polynomial/arithmetic/mod.rs
@@ -10,3 +10,10 @@
 /// [`HeightRef`](malachite_base::num::arithmetic::traits::HeightRef), the largest of the magnitudes
 /// of a polynomial's coefficients.
 pub mod height;
+/// An implementation of [`ModIsReduced`](malachite_base::num::arithmetic::traits::ModIsReduced),
+/// which checks whether every coefficient of a polynomial is less than a given modulus.
+pub mod mod_is_reduced;
+/// An implementation of
+/// [`ModPowerOf2IsReduced`](malachite_base::num::arithmetic::traits::ModPowerOf2IsReduced), which
+/// checks whether every coefficient of a polynomial is less than a given power of 2.
+pub mod mod_power_of_2_is_reduced;